mcp_video/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Redis and HTTP clients
redis>=5.0.1
httpx>=0.25.2

# Utility dependencies
python-multipart>=0.0.6
aiofiles>=23.2.1
pillow>=10.1.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional security dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
